split test_core.py to subfiles
