--- tests/test_core.py
@@ -1,1661 +1,0 @@
-# test_base.py
-"""Tests for rom_operator_inference._*.py."""
-
-import os
-import h5py
-import pytest
-import warnings
-import itertools
-import numpy as np
-from scipy import linalg as la
-
-import rom_operator_inference as roi
-
-
-# Helper functions for testing ================================================
-_MODEL_KEYS = roi._base._BaseROM._MODEL_KEYS
-_MODEL_FORMS = [''.join(s) for k in range(1, len(_MODEL_KEYS)+1)
-                           for s in itertools.combinations(_MODEL_KEYS, k)]
-_LSTSQ_REPORTS = ["datacond_", "dataregcond_", "residual_", "misfit_"]
-
-def _get_data(n=60, k=25, m=20):
-    """Get fake snapshot, velocity, and input data."""
-    X = np.random.random((n,k))
-    Xdot = np.zeros((n,k))
-    U = np.ones((m,k))
-
-    return X, Xdot, U
-
-def _get_operators(n=60, m=20):
-    """Construct fake model operators."""
-    c = np.random.random(n)
-    A = np.eye(n)
-    H = np.zeros((n,n**2))
-    Hc = np.zeros((n,n*(n+1)//2))
-    G = np.zeros((n,n**3))
-    Gc = np.zeros((n,n*(n+1)*(n+2)//6))
-    B = np.random.random((n,m)) if m else None
-    return c, A, H, Hc, G, Gc, B
-
-def _trainedmodel(continuous, modelform, Vr, m=20):
-    """Construct a base class with model operators already constructed."""
-    if continuous == "inferred":
-        ModelClass = roi._inferred.InferredContinuousROM
-    elif continuous:
-        ModelClass = roi._base._ContinuousROM
-    else:
-        ModelClass = roi._base._DiscreteROM
-
-    n,r = Vr.shape
-    c, A, H, Hc, G, Gc, B = _get_operators(r, m)
-    operators = {}
-    if "c" in modelform:
-        operators['c_'] = c
-    if "A" in modelform:
-        operators['A_'] = A
-    if "H" in modelform:
-        operators['Hc_'] = Hc
-    if "G" in modelform:
-        operators['Gc_'] = Gc
-    if "B" in modelform:
-        operators['B_'] = B
-
-    model = ModelClass(modelform)._set_operators(Vr, **operators)
-
-    model.datacond_ = np.random.random()
-    model.dataregcond_ = model.datacond_ / 2
-    model.residual_ = np.random.random()
-    model.misfit_ = model.residual_ / 2
-
-    return model
-
-
-# Helper functions and classes (public) =======================================
-def test_select_model_class():
-    """Test select_model_class()."""
-    # Try with bad `time` argument.
-    with pytest.raises(ValueError) as ex:
-        roi.select_model_class("semidiscrete", "inferred", False)
-    assert "input `time` must be one of " in ex.value.args[0]
-
-    # Try with bad `rom_strategy` argument.
-    with pytest.raises(ValueError) as ex:
-        roi.select_model_class("discrete", "opinf", False)
-    assert "input `rom_strategy` must be one of " in ex.value.args[0]
-
-    # Try with bad `parametric` argument.
-    with pytest.raises(ValueError) as ex:
-        roi.select_model_class("discrete", "inferred", True)
-    assert "input `parametric` must be one of " in ex.value.args[0]
-
-    # Try with bad combination.
-    with pytest.raises(NotImplementedError) as ex:
-        roi.select_model_class("discrete", "intrusive", "interpolated")
-    assert ex.value.args[0] == "model type invalid or not implemented"
-
-    # Valid cases.
-    assert roi.select_model_class("discrete", "inferred") is  \
-                                        roi.InferredDiscreteROM
-    assert roi.select_model_class("continuous", "inferred") is \
-                                        roi.InferredContinuousROM
-    assert roi.select_model_class("discrete", "intrusive") is \
-                                        roi.IntrusiveDiscreteROM
-    assert roi.select_model_class("continuous", "intrusive") is \
-                                        roi.IntrusiveContinuousROM
-    assert roi.select_model_class("discrete", "intrusive", "affine") is \
-                                        roi.AffineIntrusiveDiscreteROM
-    assert roi.select_model_class("continuous", "intrusive", "affine") is \
-                                        roi.AffineIntrusiveContinuousROM
-    assert roi.select_model_class("discrete", "inferred", "affine") is \
-                                        roi.AffineInferredDiscreteROM
-    assert roi.select_model_class("continuous", "inferred", "affine") is \
-                                        roi.AffineInferredContinuousROM
-    assert roi.select_model_class("discrete", "inferred", "interpolated") is \
-                                        roi.InterpolatedInferredDiscreteROM
-    assert roi.select_model_class("continuous", "inferred", "interpolated") is \
-                                        roi.InterpolatedInferredContinuousROM
-
-
-def test_load_model():
-    """Test load_model()."""
-    # Get test operators.
-    n, m, r = 20, 2, 5
-    Vr = np.random.random((n,r))
-    c_, A_, H_, Hc_, G_, Gc_, B_ = _get_operators(n=r, m=m)
-
-    # Try loading a file that does not exist.
-    target = "loadmodeltest.h5"
-    if os.path.isfile(target):                  # pragma: no cover
-        os.remove(target)
-    with pytest.raises(FileNotFoundError) as ex:
-        model = roi.load_model(target)
-    assert ex.value.args[0] == target
-
-    # Make an empty HDF5 file to start with.
-    with h5py.File(target, 'w') as f:
-        pass
-
-    with pytest.raises(ValueError) as ex:
-        model = roi.load_model(target)
-    assert ex.value.args[0] == "invalid save format (meta/ not found)"
-
-    # Make a (mostly) compatible HDF5 file to start with.
-    with h5py.File(target, 'a') as f:
-        # Store metadata.
-        meta = f.create_dataset("meta", shape=(0,))
-        meta.attrs["modelclass"] = "InferredDiscreteROOM"
-        meta.attrs["modelform"] = "cAB"
-
-    with pytest.raises(ValueError) as ex:
-        model = roi.load_model(target)
-    assert ex.value.args[0] == "invalid save format (operators/ not found)"
-
-    # Store the arrays.
-    with h5py.File(target, 'a') as f:
-        f.create_dataset("operators/c_", data=c_)
-        f.create_dataset("operators/A_", data=A_)
-        f.create_dataset("operators/B_", data=B_)
-
-    # Try to load the file, which has a bad modelclass attribute.
-    with pytest.raises(ValueError) as ex:
-        model = roi.load_model(target)
-    assert ex.value.args[0] == \
-        "invalid modelclass 'InferredDiscreteROOM' (meta.attrs)"
-
-    # Fix the file.
-    with h5py.File(target, 'a') as f:
-        f["meta"].attrs["modelclass"] = "InferredDiscreteROM"
-
-    def _check_model(mdl):
-        assert isinstance(mdl, roi.InferredDiscreteROM)
-        for attr in ["modelform",
-                     "n", "r", "m",
-                     "c_", "A_", "Hc_", "Gc_", "B_", "Vr"]:
-            assert hasattr(mdl, attr)
-        assert mdl.modelform == "cAB"
-        assert model.r == r
-        assert model.m == m
-        assert np.allclose(mdl.c_, c_)
-        assert np.allclose(mdl.A_, A_)
-        assert mdl.Hc_ is None
-        assert mdl.Gc_ is None
-        assert np.allclose(mdl.B_, B_)
-
-    # Load the file correctly.
-    model = roi.load_model(target)
-    _check_model(model)
-    assert model.Vr is None
-    assert model.n is None
-
-    # Add the basis and then load the file correctly.
-    with h5py.File(target, 'a') as f:
-        f.create_dataset("Vr", data=Vr)
-    model = roi.load_model(target)
-    _check_model(model)
-    assert np.allclose(model.Vr, Vr)
-    assert model.n == n
-
-    # One additional test to cover other cases.
-    with h5py.File(target, 'a') as f:
-        f["meta"].attrs["modelclass"] = "InferredContinuousROM"
-        f["meta"].attrs["modelform"] = "HG"
-        f.create_dataset("operators/Hc_", data=Hc_)
-        f.create_dataset("operators/Gc_", data=Gc_)
-
-    model = roi.load_model(target)
-    assert isinstance(model, roi.InferredContinuousROM)
-    for attr in ["modelform",
-                 "n", "r", "m",
-                 "c_", "A_", "Hc_", "Gc_", "B_", "Vr"]:
-        assert hasattr(model, attr)
-    assert model.modelform == "HG"
-    assert model.r == r
-    assert model.m is None
-    assert model.c_ is None
-    assert model.A_ is None
-    assert np.allclose(model.Hc_, Hc_)
-    assert np.allclose(model.Gc_, Gc_)
-    assert model.B_ is None
-    assert np.allclose(model.Vr, Vr)
-    assert model.n == n
-
-    # Clean up.
-    os.remove(target)
-
-
-class TestAffineOperator:
-    """Test _affine.base.AffineOperator."""
-    @staticmethod
-    def _set_up_affine_attributes(n=5):
-        fs = [np.sin, np.cos, np.exp]
-        As = list(np.random.random((3,n,n)))
-        return fs, As
-
-    def test_init(self):
-        """Test _affine.base.AffineOperator.__init__()."""
-        fs, As = self._set_up_affine_attributes()
-
-        # Try with different number of functions and matrices.
-        with pytest.raises(ValueError) as ex:
-            roi.AffineOperator(fs, As[:-1])
-        assert ex.value.args[0] == "expected 3 matrices, got 2"
-
-        # Try with matrices of different shapes.
-        with pytest.raises(ValueError) as ex:
-            roi.AffineOperator(fs, As[:-1] + [np.random.random((4,4))])
-        assert ex.value.args[0] == \
-            "affine operator matrix shapes do not match ((4, 4) != (5, 5))"
-
-        # Correct usage.
-        affop = roi.AffineOperator(fs, As)
-        affop = roi.AffineOperator(fs)
-        affop.matrices = As
-
-    def test_validate_coeffs(self):
-        """Test _affine.base.AffineOperator.validate_coeffs()."""
-        fs, As = self._set_up_affine_attributes()
-
-        # Try with non-callables.
-        affop = roi.AffineOperator(As)
-        with pytest.raises(ValueError) as ex:
-            affop.validate_coeffs(10)
-        assert ex.value.args[0] == \
-            "coefficients of affine operator must be callable functions"
-
-        # Try with vector-valued functions.
-        f1 = lambda t: np.array([t, t**2])
-        affop = roi.AffineOperator([f1, f1])
-        with pytest.raises(ValueError) as ex:
-            affop.validate_coeffs(10)
-        assert ex.value.args[0] == \
-            "coefficient functions of affine operator must return a scalar"
-
-        # Correct usage.
-        affop = roi.AffineOperator(fs, As)
-        affop.validate_coeffs(0)
-
-    def test_call(self):
-        """Test _affine.base.AffineOperator.__call__()."""
-        fs, As = self._set_up_affine_attributes()
-
-        # Try without matrices set.
-        affop = roi.AffineOperator(fs)
-        with pytest.raises(RuntimeError) as ex:
-            affop(10)
-        assert ex.value.args[0] == "component matrices not initialized!"
-
-        # Correct usage.
-        affop.matrices = As
-        Ap = affop(10)
-        assert Ap.shape == (5,5)
-        assert np.allclose(Ap, np.sin(10)*As[0] + \
-                               np.cos(10)*As[1] + np.exp(10)*As[2])
-
-    def test_eq(self):
-        """Test _affine.base.AffineOperator.__eq__()."""
-        fs, As = self._set_up_affine_attributes()
-        affop1 = roi.AffineOperator(fs[:-1])
-        affop2 = roi.AffineOperator(fs, As)
-
-        assert affop1 != 1
-        assert affop1 != affop2
-        affop1 = roi.AffineOperator(fs)
-        assert affop1 != affop2
-        affop1.matrices = As
-        assert affop1 == affop2
-
-
-# Base classes (private) ======================================================
-class TestBaseROM:
-    """Test _base._BaseROM."""
-    def test_init(self):
-        """Test _BaseROM.__init__()."""
-        with pytest.raises(TypeError) as ex:
-            roi._base._BaseROM()
-        assert ex.value.args[0] == \
-            "__init__() missing 1 required positional argument: 'modelform'"
-
-        with pytest.raises(TypeError) as ex:
-            roi._base._BaseROM("cAH", False)
-        assert ex.value.args[0] == \
-            "__init__() takes 2 positional arguments but 3 were given"
-
-        with pytest.raises(RuntimeError) as ex:
-            roi._base._BaseROM("cAH")
-        assert ex.value.args[0] == \
-            "abstract class instantiation (use _ContinuousROM or _DiscreteROM)"
-
-        model = roi._base._ContinuousROM("cA")
-        assert hasattr(model, "modelform")
-        assert hasattr(model, "_form")
-        assert hasattr(model, "has_inputs")
-        assert model.modelform == "cA"
-        assert model.has_constant is True
-        assert model.has_linear is True
-        assert model.has_quadratic is False
-        assert model.has_inputs is False
-
-        model.modelform = "BHc"
-        assert model.modelform == "cHB"
-        assert model.has_constant is True
-        assert model.has_linear is False
-        assert model.has_quadratic is True
-        assert model.has_inputs is True
-
-    def test_check_modelform(self):
-        """Test _BaseROM._check_modelform()."""
-        Vr = np.random.random((60,5))
-        m = 20
-
-        # Try with invalid modelform.
-        model = roi._base._ContinuousROM("bad_form")
-        with pytest.raises(ValueError) as ex:
-            model._check_modelform(trained=False)
-        assert ex.value.args[0] == \
-            "invalid modelform key 'b'; " \
-            f"options are {', '.join(model._MODEL_KEYS)}"
-
-        # Try with untrained model.
-        model.modelform = "cAH"
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'c_' missing; call fit() to train model"
-
-        # Try with missing attributes.
-        model = _trainedmodel(True, "cAHB", Vr, m)
-        c_ = model.c_.copy()
-        del model.c_
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'c_' missing; call fit() to train model"
-        model.c_ = c_
-
-        B_ = model.B_.copy()
-        del model.B_
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'B_' missing; call fit() to train model"
-        model.B_ = B_
-
-        # Try with incorrectly set attributes.
-        A_ = model.A_.copy()
-        model.A_ = None
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'A_' is None; call fit() to train model"
-
-        model = _trainedmodel(True, "cAB", Vr, m)
-        model.Hc_ = 1
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'Hc_' should be None; call fit() to train model"
-        model.Hc_ = None
-
-        model.modelform = "cA"
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'B_' should be None; call fit() to train model"
-
-        model = _trainedmodel(False, "cAH", Vr, None)
-        model.modelform = "cAHB"
-        with pytest.raises(AttributeError) as ex:
-            model._check_modelform(trained=True)
-        assert ex.value.args[0] == \
-            "attribute 'B_' is None; call fit() to train model"
-
-    def test_set_operators(self):
-        """Test _base._BaseROM._set_operators()."""
-        n, m, r = 60, 20, 30
-        Vr = np.random.random((n, r))
-        c, A, H, Hc, G, Gc, B = _get_operators(n=r, m=m)
-
-        # Test correct usage.
-        model = roi._base._ContinuousROM("cAH")._set_operators(Vr=Vr, A_=A,
-                                                               Hc_=Hc, c_=c)
-        assert isinstance(model, roi._base._ContinuousROM)
-        assert model.modelform == "cAH"
-        assert model.n == n
-        assert model.r == r
-        assert model.m is None
-        assert np.allclose(model.Vr, Vr)
-        assert np.allclose(model.c_, c)
-        assert np.allclose(model.A_, A)
-        assert np.allclose(model.Hc_, Hc)
-        assert model.B_ is None
-        assert model.Gc_ is None
-
-        model = roi._base._DiscreteROM("GB")._set_operators(None, Gc_=Gc, B_=B)
-        assert isinstance(model, roi._base._DiscreteROM)
-        assert model.modelform == "GB"
-        assert model.n is None
-        assert model.r == r
-        assert model.m == m
-        assert model.Vr is None
-        assert model.c_ is None
-        assert model.A_ is None
-        assert model.Hc_ is None
-        assert np.allclose(model.Gc_, Gc)
-        assert np.allclose(model.B_, B)
-
-    def test_check_inputargs(self):
-        """Test _BaseROM._check_inputargs()."""
-
-        # Try with has_inputs = True but without inputs.
-        model = roi._base._DiscreteROM("cB")
-        with pytest.raises(ValueError) as ex:
-            model._check_inputargs(None, 'U')
-        assert ex.value.args[0] == \
-            "argument 'U' required since 'B' in modelform"
-
-        # Try with has_inputs = False but with inputs.
-        model.modelform = "cA"
-        with pytest.raises(ValueError) as ex:
-            model._check_inputargs(1, 'u')
-        assert ex.value.args[0] == \
-            "argument 'u' invalid since 'B' in modelform"
-
-    def test_project(self):
-        """Test _base._BaseROM.project()."""
-        n, k, m, r = 60, 50, 20, 10
-        X, Xdot, U = _get_data(n, k, m)
-        model = roi._base._ContinuousROM("c")
-        model.n, model.r, model.m = n, r, m
-        model.Vr = la.svd(X)[0][:,:r]
-
-        with pytest.raises(ValueError) as ex:
-            model.project(X[:-1,:], 'X')
-        assert ex.value.args[0] == "X not aligned with Vr, dimension 0"
-
-        for S, label in [(X, 'X'), (Xdot, 'Xdot')]:
-            S_ = model.project(S, label)
-            assert S_.shape == (r,k)
-            S_ = model.project(model.Vr.T @ S, label)
-            assert S_.shape == (r,k)
-
-    def test_operator_norm_(self):
-        """Test _base._BaseROM.operator_norm_()"""
-        # Get test data.
-        n, k, m, r = 60, 50, 20, 10
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        model = _trainedmodel(True, "cAHGB", Vr, m)
-        O_ = np.concatenate((model.c_[:,np.newaxis], model.A_,
-                             model.Hc_, model.Gc_, model.B_), axis=1)
-        assert np.isclose(la.norm(O_, ord='fro')**2, model.operator_norm_)
-
-class TestDiscreteROM:
-    """Test _base._DiscreteROM."""
-    def test_construct_f_(self):
-        """Test _base.DiscreteROM._construct_f_()."""
-        model = roi._base._DiscreteROM('')
-
-        # Check that the constructed f takes the right number of arguments.
-        model.modelform = "cA"
-        model.c_, model.A_ = 1, 1
-        model.Hc_, model.Gc_, model.B_ = None, None, None
-        model._construct_f_()
-        with pytest.raises(TypeError) as ex:
-            model.f_(1, 2)
-        assert ex.value.args[0] == \
-            "<lambda>() takes 1 positional argument but 2 were given"
-
-        model.modelform = "HGB"
-        model.Hc_, model.Gc_, model.B_ = 1, 1, 1
-        model.c_, model.A_ = None, None
-        model._construct_f_()
-        with pytest.raises(TypeError) as ex:
-            model.f_(1)
-        assert ex.value.args[0] == \
-            "<lambda>() missing 1 required positional argument: 'u'"
-
-    def test_fit(self):
-        """Test _base._DiscreteROM.fit()."""
-        model = roi._base._DiscreteROM("A")
-        with pytest.raises(NotImplementedError) as ex:
-            model.fit()
-        assert ex.value.args[0] == \
-            "fit() must be implemented by child classes"
-
-        with pytest.raises(NotImplementedError) as ex:
-            model.fit(1, 2, 3, 4, 5, 6, 7, a=8)
-        assert ex.value.args[0] == \
-            "fit() must be implemented by child classes"
-
-    def test_predict(self):
-        """Test _base._DiscreteROM.predict()."""
-        model = roi._base._DiscreteROM('')
-
-        # Get test data.
-        n, k, m, r = 60, 50, 20, 10
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        # Get test (reduced) operators.
-        c, A, H, Hc, G, Gc, B = _get_operators(r, m)
-
-        niters = 5
-        x0 = X[:,0]
-        U = np.ones((m, niters-1))
-
-        # Try to predict with invalid initial condition.
-        x0_ = Vr.T @ x0
-        model = _trainedmodel(False, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0_[:-1], niters, U)
-        assert ex.value.args[0] == "x0 not aligned with Vr, dimension 0"
-
-        # Try to predict with bad niters argument.
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, -18, U)
-        assert ex.value.args[0] == \
-            "argument 'niters' must be a nonnegative integer"
-
-        # Try to predict with badly-shaped discrete inputs.
-        model = _trainedmodel(False, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, niters, np.random.random((m-1, niters-1)))
-        assert ex.value.args[0] == \
-            f"invalid input shape ({(m-1,niters-1)} != {(m,niters-1)}"
-
-        model = _trainedmodel(False, "cAHB", Vr, m=1)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, niters, np.random.random((2, niters-1)))
-        assert ex.value.args[0] == \
-            f"invalid input shape ({(2,niters-1)} != {(1,niters-1)}"
-
-        # Try to predict with continuous inputs.
-        model = _trainedmodel(False, "cAHB", Vr, m)
-        with pytest.raises(TypeError) as ex:
-            model.predict(x0, niters, lambda t: np.ones(m-1))
-        assert ex.value.args[0] == "input U must be an array, not a callable"
-
-        for form in _MODEL_FORMS:
-            if "B" not in form:             # No control inputs.
-                model = _trainedmodel(False, form, Vr, None)
-                out = model.predict(x0, niters)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,niters)
-            else:                           # Has Control inputs.
-                # Predict with 2D inputs.
-                model = _trainedmodel(False, form, Vr, m)
-                out = model.predict(x0, niters, U)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,niters)
-
-                # Predict with 1D inputs.
-                model = _trainedmodel(False, form, Vr, 1)
-                out = model.predict(x0, niters, np.ones(niters))
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,niters)
-
-        # Predict with no basis gives result in low-dimensional space.
-        model = _trainedmodel(False, "cA", Vr, None)
-        model.Vr, model.n = None, None
-        out = model.predict(Vr.T @ x0, niters)
-        assert isinstance(out, np.ndarray)
-        assert out.shape == (r,niters)
-
-
-class TestContinuousROM:
-    """Test _base._ContinuousROM."""
-    def test_construct_f_(self):
-        """Test incorrect usage of _base.ContinuousROM._construct_f_()."""
-        model = roi._base._ContinuousROM('')
-
-        # Check that the constructed f takes the right number of arguments.
-        model.modelform = "cA"
-        model.c_, model.A_ = 1, 1
-        model.Hc_, model.Gc_, model.B_ = None, None, None
-        model._construct_f_()
-        with pytest.raises(TypeError) as ex:
-            model.f_(1)
-        assert ex.value.args[0] == \
-            "<lambda>() missing 1 required positional argument: 'x_'"
-
-    def test_fit(self):
-        """Test _base._ContinuousROM.fit()."""
-        model = roi._base._ContinuousROM("A")
-        with pytest.raises(NotImplementedError) as ex:
-            model.fit()
-        assert ex.value.args[0] == \
-            "fit() must be implemented by child classes"
-
-        with pytest.raises(NotImplementedError) as ex:
-            model.fit(1, 2, 3, 4, 5, 6, 7, a=8)
-        assert ex.value.args[0] == \
-            "fit() must be implemented by child classes"
-
-    def test_predict(self):
-        """Test _base._ContinuousROM.predict()."""
-        model = roi._base._ContinuousROM('')
-
-        # Get test data.
-        n, k, m, r = 60, 50, 20, 10
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        # Get test (reduced) operators.
-        c, A, H, Hc, G, Gc, B = _get_operators(r, m)
-
-        nt = 5
-        x0 = X[:,0]
-        t = np.linspace(0, .01*nt, nt)
-        u = lambda t: np.ones(m)
-        Upred = np.ones((m, nt))
-
-        # Try to predict with invalid initial condition.
-        x0_ = Vr.T @ x0
-        model = _trainedmodel(True, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0_[1:], t, u)
-        assert ex.value.args[0] == "x0 not aligned with Vr, dimension 0"
-
-        # Try to predict with bad time array.
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, np.vstack((t,t)), u)
-        assert ex.value.args[0] == "time 't' must be one-dimensional"
-
-        # Predict without inputs.
-        for form in _MODEL_FORMS:
-            if "B" not in form:
-                model = _trainedmodel(True, form, Vr, None)
-                out = model.predict(x0, t)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,t.size)
-
-        # Predict with no basis gives result in low-dimensional space.
-        model = _trainedmodel(True, "cA", Vr, None)
-        model.Vr, model.n = None, None
-        out = model.predict(Vr.T @ x0, t)
-        assert isinstance(out, np.ndarray)
-        assert out.shape == (r,t.size)
-
-        # Try to predict with badly-shaped discrete inputs.
-        model = _trainedmodel(True, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, np.random.random((m-1, nt)))
-        assert ex.value.args[0] == \
-            f"invalid input shape ({(m-1,nt)} != {(m,nt)}"
-
-        model = _trainedmodel(True, "cAHB", Vr, m=1)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, np.random.random((2, nt)))
-        assert ex.value.args[0] == \
-            f"invalid input shape ({(2,nt)} != {(1,nt)}"
-
-        # Try to predict with badly-shaped continuous inputs.
-        model = _trainedmodel(True, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, lambda t: np.ones(m-1))
-        assert ex.value.args[0] == \
-            f"input function u() must return ndarray of shape (m,)={(m,)}"
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, lambda t: 1)
-        assert ex.value.args[0] == \
-            f"input function u() must return ndarray of shape (m,)={(m,)}"
-
-        model = _trainedmodel(True, "cAHB", Vr, m=1)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, u)
-        assert ex.value.args[0] == \
-            f"input function u() must return ndarray of shape (m,)={(1,)}" \
-            " or scalar"
-
-        # Try to predict with continuous inputs with bad return type
-        model = _trainedmodel(True, "cAHB", Vr, m)
-        with pytest.raises(ValueError) as ex:
-            model.predict(x0, t, lambda t: set([5]))
-        assert ex.value.args[0] == \
-            f"input function u() must return ndarray of shape (m,)={(m,)}"
-
-        for form in _MODEL_FORMS:
-            if "B" in form:
-                # Predict with 2D inputs.
-                model = _trainedmodel(True, form, Vr, m)
-                # continuous input.
-                out = model.predict(x0, t, u)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,nt)
-                # discrete input.
-                out = model.predict(x0, t, Upred)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,nt)
-
-                # Predict with 1D inputs.
-                model = _trainedmodel(True, form, Vr, 1)
-                # continuous input.
-                out = model.predict(x0, t, lambda t: 1)
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,nt)
-                out = model.predict(x0, t, lambda t: np.array([1]))
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,nt)
-                # discrete input.
-                out = model.predict(x0, t, np.ones_like(t))
-                assert isinstance(out, np.ndarray)
-                assert out.shape == (n,nt)
-
-
-# Basic mixins (private) ======================================================
-class TestInferredMixin:
-    """Test _inferred._InferredMixin."""
-    def test_check_training_data_shapes(self):
-        """Test _inferred._InferredMixin._check_training_data_shapes()."""
-        # Get test data.
-        n, k, m, r = 60, 50, 20, 10
-        X, Xdot, U = _get_data(n, k, m)
-        model = roi._inferred._InferredMixin()
-
-        # Try to fit the model with misaligned X and Xdot.
-        with pytest.raises(ValueError) as ex:
-            model._check_training_data_shapes([X, Xdot[:,1:-1]])
-        assert ex.value.args[0] == "data sets not aligned, dimension 1"
-
-        # Try to fit the model with misaligned X and U.
-        with pytest.raises(ValueError) as ex:
-            model._check_training_data_shapes([X, Xdot, U[:,:-1]])
-        assert ex.value.args[0] == "data sets not aligned, dimension 1"
-
-        model._check_training_data_shapes([X, Xdot])
-        model._check_training_data_shapes([X, Xdot, U])
-
-    def _test_fit(self, ModelClass):
-        """Test _inferred._InferredMixin.fit(), the parent method for
-        _inferred.InferredDiscreteROM.fit() and
-        _inferred.InferredContinuousROM.fit().
-        """
-        model = ModelClass("cAH")
-
-        # Get test data.
-        n, k, m, r = 60, 500, 20, 10
-        X, Xdot, U = _get_data(n, k, m)
-        Vr = la.svd(X)[0][:,:r]
-        args = [Vr, X]
-        if issubclass(ModelClass, roi._inferred._ContinuousROM):
-            args.insert(1, Xdot)
-
-        # Fit the model with each possible non-input modelform.
-        for form in _MODEL_FORMS:
-            if "B" not in form:
-                model.modelform = form
-                model.fit(*args)
-
-        def _test_output_shapes(model):
-            """Test shapes of output operators for modelform="cAHB"."""
-            assert model.r == r
-            assert model.m == m
-            assert model.c_.shape == (r,)
-            assert model.A_.shape == (r,r)
-            assert model.Hc_.shape == (r,r*(r+1)//2)
-            assert model.H_.shape == (r,r**2)
-            assert model.Gc_.shape == (r,r*(r+1)*(r+2)//6)
-            assert model.G_.shape == (r,r**3)
-            assert model.B_.shape == (r,m)
-            assert hasattr(model, "datacond_")
-            assert hasattr(model, "dataregcond_")
-            assert round(model.dataregcond_, 6) <= round(model.datacond_, 6)
-            assert hasattr(model, "residual_")
-            assert hasattr(model, "misfit_")
-            assert round(model.misfit_, 6) <= round(model.residual_, 6)
-
-        # Test with high-dimensional inputs.
-        model.modelform = "cAHGB"
-        model.fit(*args, U=U)
-        _test_output_shapes(model)
-        assert model.n == n
-        assert np.allclose(model.Vr, Vr)
-
-        # Test again with one-dimensional inputs.
-        m = 1
-        model.fit(*args, U=np.random.random(k))
-        _test_output_shapes(model)
-        assert model.n == n
-        assert np.allclose(model.Vr, Vr)
-
-        # Test again with Vr = None and projected data.
-        args[0] = None
-        for i in range(1,len(args)):
-            args[i] = Vr.T @ args[i]
-        model.fit(*args, U=np.random.random(k))
-        _test_output_shapes(model)
-        assert model.n is None
-        assert model.Vr is None
-
-
-class TestIntrusiveMixin:
-    """Test _intrusive._IntrusiveMixin."""
-    def test_check_operators(self):
-        """Test _intrusive._IntrusiveMixin._check_operators()."""
-        model = roi._intrusive._IntrusiveMixin()
-        model.modelform = "cAHB"
-        v = None
-
-        # Try with missing operator keys.
-        with pytest.raises(KeyError) as ex:
-            model._check_operators({"A":v, "H":v, "B":v})
-        assert ex.value.args[0] == "missing operator key 'c'"
-
-        with pytest.raises(KeyError) as ex:
-            model._check_operators({"H":v, "B":v})
-        assert ex.value.args[0] == "missing operator keys 'c', 'A'"
-
-        # Try with surplus operator keys.
-        with pytest.raises(KeyError) as ex:
-            model._check_operators({'CC':v, "c":v, "A":v, "H":v, "B":v})
-        assert ex.value.args[0] == "invalid operator key 'CC'"
-
-        with pytest.raises(KeyError) as ex:
-            model._check_operators({"c":v, "A":v, "H":v, "B":v,
-                                    'CC':v, 'LL':v})
-        assert ex.value.args[0] == "invalid operator keys 'CC', 'LL'"
-
-        # Correct usage.
-        model._check_operators({"c":v, "A":v, "H":v, "B":v})
-
-    def _test_fit(self, ModelClass):
-        """Test _intrusive._IntrusiveMixin.fit(), the parent method for
-        _intrusive.IntrusiveDiscreteROM.fit() and
-        _intrusive.IntrusiveContinuousROM.fit().
-        """
-        model = ModelClass("cAHB")
-
-        # Get test data.
-        n, k, m, r = 30, 50, 10, 5
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        # Get test operators.
-        c, A, H, Hc, G, Gc, B = _get_operators(n, m)
-        B1d = B[:,0]
-        operators = {"c":c, "A":A, "H":H, "G":G, "B":B}
-
-        # Try to fit the model with misaligned operators and Vr.
-        Abad = A[:,:-2]
-        Hbad = H[:,1:]
-        Gbad = G[:,:-1]
-        cbad = c[::2]
-        Bbad = B[1:,:]
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {"c":cbad, "A":A, "H":H, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator c not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {"c":c, "A":Abad, "H":H, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator A not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {"c":c, "A":A, "H":Hbad, "B":B})
-        assert ex.value.args[0] == \
-            "basis Vr and FOM operator H not aligned"
-
-        model = ModelClass("cAGB")
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {"c":c, "A":A, "G":Gbad, "B":B})
-        assert ex.value.args[0] == \
-            "basis Vr and FOM operator G not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {"c":c, "A":A, "G":G, "B":Bbad})
-        assert ex.value.args[0] == "basis Vr and FOM operator B not aligned"
-
-        # Fit the model with each possible modelform.
-        for form in ["A", "cA", "H", "cH", "AG", "cAH", "cAHB"]:
-            model.modelform = form
-            ops = {key:val for key,val in operators.items() if key in form}
-            model.fit(Vr, ops)
-
-        model.modelform = "cAHGB"
-        model.fit(Vr, {"c":c, "A":A, "H":Hc, "G":Gc, "B":B})
-
-        # Test fit output sizes.
-        assert model.n == n
-        assert model.r == r
-        assert model.m == m
-        assert model.A.shape == (n,n)
-        assert model.Hc.shape == (n,n*(n+1)//2)
-        assert model.H.shape == (n,n**2)
-        assert model.Gc.shape == (n,n*(n+1)*(n+2)//6)
-        assert model.G.shape == (n,n**3)
-        assert model.c.shape == (n,)
-        assert model.B.shape == (n,m)
-        assert model.A_.shape == (r,r)
-        assert model.Hc_.shape == (r,r*(r+1)//2)
-        assert model.H_.shape == (r,r**2)
-        assert model.Gc_.shape == (r,r*(r+1)*(r+2)//6)
-        assert model.G_.shape == (r,r**3)
-        assert model.c_.shape == (r,)
-        assert model.B_.shape == (r,m)
-
-        # Fit the model with 1D inputs (1D array for B)
-        model.modelform = "cAHB"
-        model.fit(Vr, {"c":c, "A":A, "H":H, "B":B1d})
-        assert model.B.shape == (n,1)
-        assert model.B_.shape == (r,1)
-
-
-class TestNonparametricMixin:
-    """Test _base._NonparametricMixin."""
-    def test_str(self):
-        """Test _base._NonparametricMixin.__str__() (string representation)."""
-        # Continuous ROMs
-        model = roi._inferred.InferredContinuousROM("A")
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = Ax(t)"
-        model.modelform = "cA"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = c + Ax(t)"
-        model.modelform = "HB"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = H(x(t) ⊗ x(t)) + Bu(t)"
-        model.modelform = "G"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = G(x(t) ⊗ x(t) ⊗ x(t))"
-        model.modelform = "cH"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = c + H(x(t) ⊗ x(t))"
-
-        # Discrete ROMs
-        model = roi._intrusive.IntrusiveDiscreteROM("A")
-        assert str(model) == \
-            "Reduced-order model structure: x_{j+1} = Ax_{j}"
-        model.modelform = "cB"
-        assert str(model) == \
-            "Reduced-order model structure: x_{j+1} = c + Bu_{j}"
-        model.modelform = "H"
-        assert str(model) == \
-            "Reduced-order model structure: x_{j+1} = H(x_{j} ⊗ x_{j})"
-
-    def test_save_model(self):
-        """Test _base._NonparametricMixin.save_model()."""
-        # Clean up after old tests.
-        target = "savemodeltest.h5"
-        if os.path.isfile(target):              # pragma: no cover
-            os.remove(target)
-
-        # Get a test model.
-        n, m, r = 15, 2, 5
-        Vr = np.random.random((n,r))
-        model = _trainedmodel("inferred", "cAHGB", Vr, m)
-
-        def _checkfile(filename, mdl, hasbasis):
-            assert os.path.isfile(filename)
-            with h5py.File(filename, 'r') as data:
-                # Check metadata.
-                assert "meta" in data
-                assert len(data["meta"]) == 0
-                assert data["meta"].attrs["modelclass"] == \
-                                                    mdl.__class__.__name__
-                assert data["meta"].attrs["modelform"] == mdl.modelform
-
-                # Check basis
-                if hasbasis:
-                    assert "Vr" in data
-                    assert np.allclose(data["Vr"], Vr)
-
-                # Check operators
-                assert "operators" in data
-                if "c" in mdl.modelform:
-                    assert np.allclose(data["operators/c_"], mdl.c_)
-                else:
-                    assert "c_" not in data["operators"]
-                if "A" in mdl.modelform:
-                    assert np.allclose(data["operators/A_"], mdl.A_)
-                else:
-                    assert "A_" not in data["operators"]
-                if "H" in mdl.modelform:
-                    assert np.allclose(data["operators/Hc_"], mdl.Hc_)
-                else:
-                    assert "Hc_" not in data["operators"]
-                if "G" in mdl.modelform:
-                    assert np.allclose(data["operators/Gc_"], mdl.Gc_)
-                else:
-                    assert "Gc_" not in data["operators"]
-                if "B" in mdl.modelform:
-                    assert np.allclose(data["operators/B_"], mdl.B_)
-                else:
-                    assert "B_" not in data["operators"]
-
-                # Check other attributes.
-                assert "other" in data
-                for attr in _LSTSQ_REPORTS:
-                    assert data[f"other/{attr}"][0] == getattr(mdl, attr)
-
-        model.save_model(target[:-3], save_basis=False)
-        _checkfile(target, model, False)
-
-        with pytest.raises(FileExistsError) as ex:
-            model.save_model(target, overwrite=False)
-        assert ex.value.args[0] == target
-
-        model.save_model(target, save_basis=True, overwrite=True)
-        _checkfile(target, model, True)
-
-        model = _trainedmodel("inferred", "c", Vr, 0)
-        model.save_model(target, overwrite=True)
-        _checkfile(target, model, True)
-
-        model = _trainedmodel("inferred", "AB", Vr, m)
-        model.Vr = None
-        model.save_model(target, save_basis=True, overwrite=True)
-        _checkfile(target, model, False)
-
-        # Check that save_model() and load_model() are inverses.
-        model.Vr = Vr
-        model.save_model(target, save_basis=True, overwrite=True)
-        model2 = roi.load_model(target)
-        for attr in ["n", "m", "r", "modelform", "__class__"] + _LSTSQ_REPORTS:
-            assert getattr(model, attr) == getattr(model2, attr)
-        for attr in ["A_", "B_", "Vr"]:
-            assert np.allclose(getattr(model, attr), getattr(model2, attr))
-        for attr in ["c_", "Hc_", "Gc_"]:
-            assert getattr(model, attr) is getattr(model2, attr) is None
-
-        # Check Vr = None functionality.
-        model.Vr, model.n = None, None
-        model.save_model(target, overwrite=True)
-        model2 = roi.load_model(target)
-        for attr in ["m", "r", "modelform", "__class__"] + _LSTSQ_REPORTS:
-            assert getattr(model, attr) == getattr(model2, attr)
-        for attr in ["A_", "B_",]:
-            assert np.allclose(getattr(model, attr), getattr(model2, attr))
-        for attr in ["n", "c_", "Hc_", "Gc_", "Vr"]:
-            assert getattr(model, attr) is getattr(model2, attr) is None
-
-        # Try to save a bad model.
-        A_ = model.A_
-        del model.A_
-        with pytest.raises(AttributeError) as ex:
-            model.save_model(target, overwrite=True)
-
-        os.remove(target)
-
-
-class TestParametricMixin:
-    """Test _base._ParametricMixin."""
-    def test_str(self):
-        """Test _base._ParametricMixin.__str__() (string representation)."""
-        # Continuous ROMs
-        model = roi._interpolate.InterpolatedInferredContinuousROM("A")
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = Ax(t)"
-        model.c_ = lambda t: t
-        model.A_ = lambda t: t
-        model.modelform = "cA"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = c(µ) + A(µ)x(t)"
-        model.Hc_ = None
-        model.Gc_ = lambda t: t
-        model.B_ = None
-        model.modelform = "HB"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = H(x(t) ⊗ x(t)) + Bu(t)"
-        model.modelform = "G"
-        assert str(model) == \
-            "Reduced-order model structure: dx / dt = G(µ)(x(t) ⊗ x(t) ⊗ x(t))"
-
-        # Discrete ROMs
-        model = roi.AffineIntrusiveDiscreteROM("cH")
-        assert str(model) == \
-            "Reduced-order model structure: x_{j+1} = c + H(x_{j} ⊗ x_{j})"
-        model.c_ = lambda t: t
-        model.Hc_ = None
-        assert str(model) == \
-            "Reduced-order model structure: x_{j+1} = c(µ) + H(x_{j} ⊗ x_{j})"
-
-# Specialized mixins (private) ================================================
-class TestInterpolatedInferredMixin:
-    """Test _interpolate.inferred._InterpolatedInferredMixin."""
-    pass
-
-
-class TestAffineMixin:
-    """Test _affine.base._AffineMixin."""
-    def test_check_affines(self):
-        """Test _affine.base._AffineMixin._check_affines()."""
-        model = roi._affine.base._AffineMixin()
-        model.modelform = "cAHB"
-        v = [lambda s: 0, lambda s: 0]
-
-        # Try with surplus affine keys.
-        with pytest.raises(KeyError) as ex:
-            model._check_affines({'CC':v, "c":v, "A":v, "H":v, "B":v}, 0)
-        assert ex.value.args[0] == "invalid affine key 'CC'"
-
-        with pytest.raises(KeyError) as ex:
-            model._check_affines({"c":v, "A":v, "H":v, "B":v,
-                                    'CC':v, 'LL':v}, 0)
-        assert ex.value.args[0] == "invalid affine keys 'CC', 'LL'"
-
-        # Correct usage.
-        model._check_affines({"c":v, "H":v}, 0)     # OK to be missing some.
-        model._check_affines({"c":v, "A":v, "H":v, "B":v}, 0)
-
-    def _test_predict(self, ModelClass):
-        """Test predict() methods for Affine classes:
-        * _affine.inferred.AffineInferredDiscreteROM.predict()
-        * _affine.inferred.AffineInferredContinuousROM.predict()
-        * _affine.intrusive.AffineIntrusiveDiscreteROM.predict()
-        * _affine.intrusive.AffineIntrusiveContinuousROM.predict()
-        """
-        model = ModelClass("cAHG")
-
-        # Get test data.
-        n, k, m, r = 60, 50, 20, 10
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        # Get test operators.
-        ident = lambda a: a
-        c, A, H, Hc, G, Gc, B = _get_operators(r, m)
-        model.Vr = Vr
-        model.c_ = roi.AffineOperator([ident, ident], [c,c])
-        model.A_ = roi.AffineOperator([ident, ident, ident], [A,A,A])
-        model.Hc_ = roi.AffineOperator([ident], [Hc])
-        model.Gc_ = roi.AffineOperator([ident, ident], [Gc, Gc])
-        model.B_ = None
-
-        # Predict.
-        if issubclass(ModelClass, roi._base._ContinuousROM):
-            model.predict(1, X[:,0], np.linspace(0, 1, 100))
-        else:
-            model.predict(1, X[:,0], 100)
-
-
-class TestAffineInferredMixin:
-    """Test _affine.inferred._AffineInferredMixin."""
-    def _test_fit(self, ModelClass):
-        """Test _affine.inferred._AffineInferredMixin.fit(), parent method of
-        _affine.inferred.AffineInferredDiscreteROM.fit() and
-        _affine.inferred.AffineInferredContinuousROM.fit().
-        """
-        model = ModelClass("cAHB")
-        is_continuous = issubclass(ModelClass, roi._base._ContinuousROM)
-
-        # Get test data.
-        n, k, m, r, s = 50, 200, 5, 10, 3
-        X, Xdot, U = _get_data(n, k, m)
-        Vr = la.svd(X)[0][:,:r]
-        θs = [lambda µ: µ, lambda µ: µ**2, lambda µ: µ**3, lambda µ: µ**4]
-        µs = np.arange(1, s+1)
-        affines = {"c": θs[:2],
-                   "A": θs,
-                   "H": θs[:1],
-                   "G": θs[:1],
-                   "B": θs[:3]}
-        Xs, Xdots, Us = [X]*s, [Xdot]*s, [U]*s
-        args = [Vr, µs, affines, Xs]
-        if is_continuous:
-            args.insert(3, Xdots)
-
-        # Try with bad number of parameters.
-        model.modelform = "cAHGB"
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, µs[:-1], *args[2:], Us)
-        assert ex.value.args[0] == \
-            f"num parameter samples != num state snapshot sets ({s-1} != {s})"
-
-        if is_continuous:
-            with pytest.raises(ValueError) as ex:
-                model.fit(Vr, µs, affines, Xs, Xdots[:-1], Us)
-            assert ex.value.args[0] == \
-                f"num parameter samples != num rhs sets ({s} != {s-1})"
-
-        # Try with varying input sizes.
-        if is_continuous:
-            with pytest.raises(ValueError) as ex:
-                model.fit(Vr, µs, affines, Xs, Xdots, [U, U[:-1], U])
-            assert ex.value.args[0] == "control inputs not aligned"
-        else:
-            with pytest.raises(ValueError) as ex:
-                model.fit(Vr, µs, affines, Xs, [U, U[:-1], U])
-            assert ex.value.args[0] == "control inputs not aligned"
-
-        for form in _MODEL_FORMS:
-            args[2] = {key:val for key,val in affines.items() if key in form}
-            model.modelform = form
-            model.fit(*args, Us=Us if "B" in form else None)
-
-            args[2] = {} # Non-affine case.
-            model.fit(*args, Us=Us if "B" in form else None)
-
-        def _test_output_shapes(model):
-            """Test shapes of output operators for modelform="cAHB"."""
-            assert model.n == n
-            assert model.r == r
-            assert model.m == m
-            assert model.A_.shape == (r,r)
-            assert model.Hc_.shape == (r,r*(r+1)//2)
-            assert model.H_.shape == (r,r**2)
-            assert model.c_.shape == (r,)
-            assert model.B_.shape == (r,m)
-            assert hasattr(model, "residual_")
-
-        model.modelform = "cAHB"
-        model.fit(*args, Us=Us)
-        _test_output_shapes(model)
-
-        # Fit the model with 1D inputs (1D array for B)
-        model.modelform = "cAHB"
-        model.fit(*args, Us=np.ones((s,k)))
-        m = 1
-        _test_output_shapes(model)
-
-
-class TestAffineIntrusiveMixin:
-    """Test _affine.intrusive._AffineIntrusiveMixin."""
-    def _test_fit(self, ModelClass):
-        """Test _affine.intrusive._AffineIntrusiveMixin.fit(), parent method of
-        _affine.intrusive.AffineIntrusiveDiscreteROM.fit() and
-        _affine.intrusive.AffineIntrusiveContinuousROM.fit().
-        """
-        model = ModelClass("cAHGB")
-
-        # Get test data.
-        n, k, m, r = 30, 1000, 10, 5
-        X = _get_data(n, k, m)[0]
-        Vr = la.svd(X)[0][:,:r]
-
-        # Get test operators.
-        c, A, H, Hc, G, Gc, B = _get_operators(n, m)
-        B1d = B[:,0]
-        ident = lambda a: a
-        affines = {"c": [ident, ident],
-                   "A": [ident, ident, ident],
-                   "H": [ident],
-                   "G": [ident],
-                   "B": [ident, ident]}
-        operators = {"c": [c, c],
-                     "A": [A, A, A],
-                     "H": [H],
-                     "G": [G],
-                     "B": [B, B]}
-
-        # Try to fit the model with misaligned operators and Vr.
-        Abad = A[:,:-2]
-        Hbad = H[:,1:]
-        Gbad = G[:,:-1]
-        cbad = c[::2]
-        Bbad = B[1:,:]
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, affines,
-                      {"c":[cbad, cbad],
-                       "A": [A, A, A],
-                       "H": [H],
-                       "G": [G],
-                       "B": [B, B]})
-        assert ex.value.args[0] == "basis Vr and FOM operator c not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, affines,
-                      {"c":[c, c],
-                       "A": [Abad, Abad, Abad],
-                       "H": [H],
-                       "G": [G],
-                       "B": [B, B]})
-        assert ex.value.args[0] == "basis Vr and FOM operator A not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, affines,
-                      {"c":[c, c],
-                       "A": [A, A, A],
-                       "H": [Hbad],
-                       "G": [G],
-                       "B": [B, B]})
-        assert ex.value.args[0] == \
-            "basis Vr and FOM operator H not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, affines,
-                      {"c":[c, c],
-                       "A": [A, A, A],
-                       "H": [H],
-                       "G": [Gbad],
-                       "B": [B, B]})
-        assert ex.value.args[0] == \
-            "basis Vr and FOM operator G not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, affines,
-                      {"c":[c, c],
-                       "A": [A, A, A],
-                       "H": [H],
-                       "G": [G],
-                       "B": [Bbad, Bbad]})
-        assert ex.value.args[0] == "basis Vr and FOM operator B not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {}, {"c":cbad, "A":A, "H":H, "G":G, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator c not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {}, {"c":c, "A":Abad, "H":H, "G":G, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator A not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {}, {"c":c, "A":A, "H":Hbad, "G":G, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator H not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {}, {"c":c, "A":A, "H":H, "G":Gbad, "B":B})
-        assert ex.value.args[0] == "basis Vr and FOM operator G not aligned"
-
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, {}, {"c":c, "A":A, "H":H, "G":G, "B":Bbad})
-        assert ex.value.args[0] == "basis Vr and FOM operator B not aligned"
-
-        # Fit the model correctly with each possible modelform.
-        for form in ["A", "cA", "H", "cH", "AG", "cAH", "cAHB"]:
-            model.modelform = form
-            afs = {key:val for key,val in affines.items() if key in form}
-            ops = {key:val for key,val in operators.items() if key in form}
-            model.fit(Vr, afs, ops)
-
-        model.modelform = "cAHGB"
-        model.fit(Vr, {}, {"c":c, "A":A, "H":H, "G":Gc, "B":B})
-        model.fit(Vr, {}, {"c":c, "A":A, "H":Hc, "G":G, "B":B})
-        model.fit(Vr, {}, {"c":c, "A":A, "H":H, "G":Gc, "B":B1d})
-        model.fit(Vr, affines,
-                  {"c":[c, c],
-                   "A": [A, A, A],
-                   "H": [Hc],
-                   "G": [Gc],
-                   "B": [B, B]})
-        model.fit(Vr, affines, operators)
-
-        # Test fit output sizes.
-        assert model.n == n
-        assert model.r == r
-        assert model.m == m
-        assert model.A.shape == (n,n)
-        assert model.Hc.shape == (n,n*(n+1)//2)
-        assert model.H.shape == (n,n**2)
-        assert model.Gc.shape == (n,n*(n+1)*(n+2)//6)
-        assert model.G.shape == (n,n**3)
-        assert model.c.shape == (n,)
-        assert model.B.shape == (n,m)
-        assert model.A_.shape == (r,r)
-        assert model.Hc_.shape == (r,r*(r+1)//2)
-        assert model.H_.shape == (r,r**2)
-        assert model.Gc_.shape == (r,r*(r+1)*(r+2)//6)
-        assert model.G_.shape == (r,r**3)
-        assert model.c_.shape == (r,)
-        assert model.B_.shape == (r,m)
-
-        # Fit the model with 1D inputs (1D array for B)
-        model.modelform = "cAHGB"
-        model.fit(Vr, affines,
-                  {"c":[c, c],
-                   "A": [A, A, A],
-                   "H": [Hc],
-                   "G": [Gc],
-                   "B": [B1d, B1d]})
-        assert model.B.shape == (n,1)
-        assert model.B_.shape == (r,1)
-
-
-# Useable classes (public) ====================================================
-# Nonparametric operator inference models -------------------------------------
-class TestInferredDiscreteROM:
-    """Test _inferred.InferredDiscreteROM."""
-    def test_fit(self):
-        """Test _inferred.InferredDiscreteROM.fit()."""
-        TestInferredMixin()._test_fit(roi.InferredDiscreteROM)
-
-
-class TestInferredContinuousROM:
-    """Test _inferred.InferredContinuousROM."""
-    def test_fit(self):
-        """Test _inferred.InferredContinuousROM.fit()."""
-        TestInferredMixin()._test_fit(roi.InferredContinuousROM)
-
-
-# Nonparametric intrusive models ----------------------------------------------
-class TestIntrusiveDiscreteROM:
-    """Test _intrusive.IntrusiveDiscreteROM."""
-    def test_fit(self):
-        """Test _intrusive.IntrusiveDiscreteROM.fit()."""
-        TestIntrusiveMixin()._test_fit(roi.IntrusiveDiscreteROM)
-
-
-class TestIntrusiveContinuousROM:
-    """Test _intrusive.IntrusiveContinuousROM."""
-    def test_fit(self):
-        """Test _intrusive.IntrusiveContinuousROM.fit()."""
-        TestIntrusiveMixin()._test_fit(roi.IntrusiveContinuousROM)
-
-
-# Interpolated operator inference models --------------------------------------
-class TestInterpolatedInferredDiscreteROM:
-    """Test _interpolate.inferred.InterpolatedInferredDiscreteROM."""
-    def test_fit(self):
-        """Test
-        _interpolate.inferred.InterpolatedInferredDiscreteROM.fit().
-        """
-        model = roi.InterpolatedInferredDiscreteROM("cAH")
-
-        # Get data for fitting.
-        n, m, k, r = 50, 10, 100, 5
-        X1, _, U1 = _get_data(n, k, m)
-        X2, U2 = X1+1, U1+1
-        Xs = [X1, X2]
-        Us = [U1, U2]
-        ps = [1, 2]
-        Vr = la.svd(np.hstack(Xs))[0][:,:r]
-
-        # Try with non-scalar parameters.
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, [np.array([1,1]), np.array([2,2])], Xs)
-        assert ex.value.args[0] == "only scalar parameter values are supported"
-
-        # Try with bad number of Xs.
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, ps, [X1, X2, X2+1])
-        assert ex.value.args[0] == \
-            "num parameter samples != num state snapshot sets (2 != 3)"
-
-        # Try with varying input sizes.
-        model.modelform = "cAHB"
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, ps, Xs, [U1, U2[:-1]])
-        assert ex.value.args[0] == "control inputs not aligned"
-
-        # Fit correctly with no inputs.
-        model.modelform = "cAH"
-        model.fit(Vr, ps, Xs)
-        for attr in ["models_", "fs_"] + [s[:-1]+"s_" for s in _LSTSQ_REPORTS]:
-            assert hasattr(model, attr)
-            assert len(getattr(model, attr)) == len(model.models_)
-
-        # Fit correctly with inputs.
-        model.modelform = "cAHGB"
-        model.fit(Vr, ps, Xs, Us)
-
-        assert len(model) == len(ps)
-
-        # Test again with Vr = None and projected inputs.
-        Xs_ = [Vr.T @ X for X in Xs]
-        model.fit(None, ps, Xs_, Us)
-        assert len(model) == len(ps)
-        assert model.Vr is None
-        assert model.n is None
-
-    def test_predict(self):
-        """Test
-        _interpolate.inferred.InterpolatedInferredDiscreteROM.predict().
-        """
-        model = roi.InterpolatedInferredDiscreteROM("cAH")
-
-        # Get data for fitting.
-        n, m, k, r = 50, 10, 100, 5
-        X1, _, U1 = _get_data(n, k, m)
-        X2, U2 = X1+1, U1+1
-        Xs = [X1, X2]
-        Us = [U1, U2]
-        ps = [1, 2]
-        Vr = la.svd(np.hstack(Xs))[0][:,:r]
-
-        # Parameters for predicting.
-        x0 = np.random.random(n)
-        niters = 5
-        U = np.ones((m,niters))
-
-        # Fit / predict with no inputs.
-        model.fit(Vr, ps, Xs)
-        model.predict(1, x0, niters)
-        model.predict(1.5, x0, niters)
-
-        # Fit / predict with inputs.
-        model.modelform = "cAHB"
-        model.fit(Vr, ps, Xs, Us)
-        model.predict(1, x0, niters, U)
-        model.predict(1.5, x0, niters, U)
-
-
-class TestInterpolatedInferredContinuousROM:
-    """Test _interpolate.inferred.InterpolatedInferredContinuousROM."""
-    def test_fit(self):
-        """Test
-        _interpolate.inferred.InterpolatedInferredContinuousROM.fit().
-        """
-        model = roi.InterpolatedInferredContinuousROM("cAH")
-
-        # Get data for fitting.
-        n, m, k, r = 50, 10, 100, 5
-        X1, Xdot1, U1 = _get_data(n, k, m)
-        X2, Xdot2, U2 = X1+1, Xdot1.copy(), U1+1
-        Xs = [X1, X2]
-        Xdots = [Xdot1, Xdot2]
-        Us = [U1, U2]
-        ps = [1, 2]
-        Vr = la.svd(np.hstack(Xs))[0][:,:r]
-
-        # Try with non-scalar parameters.
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, [np.array([1,1]), np.array([2,2])], Xs, Xdots)
-        assert ex.value.args[0] == "only scalar parameter values are supported"
-
-        # Try with bad number of Xs.
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, ps, [X1, X2, X2+1], Xdots)
-        assert ex.value.args[0] == \
-            "num parameter samples != num state snapshot sets (2 != 3)"
-
-        # Try with bad number of Xdots.
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, ps, Xs, Xdots + [Xdot1])
-        assert ex.value.args[0] == \
-            "num parameter samples != num velocity snapshot sets (2 != 3)"
-
-        # Try with varying input sizes.
-        model.modelform = "cAHB"
-        with pytest.raises(ValueError) as ex:
-            model.fit(Vr, ps, Xs, Xdots, [U1, U2[:-1]])
-        assert ex.value.args[0] == "control inputs not aligned"
-
-        # Fit correctly with no inputs.
-        model.modelform = "cAH"
-        model.fit(Vr, ps, Xs, Xdots)
-        for attr in ["models_", "fs_"] + [s[:-1]+"s_" for s in _LSTSQ_REPORTS]:
-            assert hasattr(model, attr)
-            assert len(getattr(model, attr)) == len(model.models_)
-
-        # Fit correctly with inputs.
-        model.modelform = "cAHB"
-        model.fit(Vr, ps, Xs, Xdots, Us)
-        assert len(model) == len(ps)
-
-        # Test again with Vr = None and projected inputs.
-        Xs_ = [Vr.T @ X for X in Xs]
-        Xdots_ = [Vr.T @ Xdot for Xdot in Xdots]
-        model.fit(None, ps, Xs_, Xdots_, Us)
-        assert len(model) == len(ps)
-        assert model.Vr is None
-        assert model.n is None
-
-    def test_predict(self):
-        """Test
-        _interpolate.inferred.InterpolatedInferredContinuousROM.predict().
-        """
-        model = roi.InterpolatedInferredContinuousROM("cAH")
-
-        # Get data for fitting.
-        n, m, k, r = 50, 10, 100, 5
-        X1, Xdot1, U1 = _get_data(n, k, m)
-        X2, Xdot2, U2 = X1+1, Xdot1.copy(), U1+1
-        Xs = [X1, X2]
-        Xdots = [Xdot1, Xdot2]
-        Us = [U1, U2]
-        ps = [1, 2]
-        Vr = la.svd(np.hstack(Xs))[0][:,:r]
-
-        # Parameters for predicting.
-        x0 = np.random.random(n)
-        nt = 5
-        t = np.linspace(0, .01*nt, nt)
-        u = lambda t: np.ones(10)
-
-        # Fit / predict with no inputs.
-        model.fit(Vr, ps, Xs, Xdots)
-        model.predict(1, x0, t)
-        model.predict(1.5, x0, t)
-
-        # Fit / predict with inputs.
-        model.modelform = "cAHB"
-        model.fit(Vr, ps, Xs, Xdots, Us)
-        model.predict(1, x0, t, u)
-        model.predict(1.5, x0, t, u)
-
-
-# Affine inferred models ------------------------------------------------------
-class TestAffineInferredDiscreteROM:
-    """Test _affine.inferred.AffineInferredDiscreteROM."""
-    def test_fit(self):
-        """Test _affine.inferred.AffineInferredDiscreteROM.fit()."""
-        TestAffineInferredMixin()._test_fit(roi.AffineInferredDiscreteROM)
-
-    def test_predict(self):
-        """Test _affine.inferred.AffineInferredDiscreteROM.predict()."""
-        TestAffineMixin()._test_predict(roi.AffineInferredDiscreteROM)
-
-
-class TestAffineInferredContinuousROM:
-    """Test _affine.inferred.AffineInferredContinuousROM."""
-    def test_fit(self):
-        """Test _affine.inferred.AffineInferredContinuousROM.fit()."""
-        TestAffineInferredMixin()._test_fit(roi.AffineInferredContinuousROM)
-
-    def test_predict(self):
-        """Test _affine.inferred.AffineInferredContinuousROM.predict()."""
-        TestAffineMixin()._test_predict(roi.AffineInferredContinuousROM)
-
-
-# Affine intrusive models -----------------------------------------------------
-class TestAffineIntrusiveDiscreteROM:
-    """Test _affine.intrusive.AffineIntrusiveDiscreteROM."""
-    def test_fit(self):
-        """Test _affine.intrusive.AffineIntrusiveDiscreteROM.fit()."""
-        TestAffineIntrusiveMixin()._test_fit(roi.AffineIntrusiveDiscreteROM)
-
-    def test_predict(self):
-        """Test _affine.intrusive.AffineIntrusiveDiscreteROM.predict()."""
-        TestAffineMixin()._test_predict(roi.AffineIntrusiveDiscreteROM)
-
-
-class TestAffineIntrusiveContinuousROM:
-    """Test _affine.intrusive.AffineIntrusiveContinuousROM."""
-    def test_fit(self):
-        """Test _affine.intrusive.AffineIntrusiveContinuousROM.fit()."""
-        TestAffineIntrusiveMixin()._test_fit(roi.AffineIntrusiveContinuousROM)
-
-    def test_predict(self):
-        """Test _affine.intrusive.AffineIntrusiveContinuousROM.predict()."""
-        TestAffineMixin()._test_predict(roi.AffineIntrusiveContinuousROM)
